Provide an idiomatic, higher-level "wpe" module using FFI metatypes
The ffi.metatype() function allows attaching additional behaviour to
values of C types, in a fashion similar to that of regular Lua
metatables. This is used to configure how values are created when using
direct instantiation of types with the __call metamethod. The __index
metatable entry is set to the methods that are exposed as higher level
API, and they can be used with the familiar colon-syntax for method
invocation e.g. view:get_page():load_url(...). Additionally, __gc
metamethods are configured to inform the FFI module that WKRelease()
should be called on values when they are destroyed by the GC on the Lua
side of the world.  This avoids manually having to call it to avoid
memory leaks.